requirements.txt
bs4
requests
gTTS
pybars3
rich
python-dotenv